Complete migration of cargo subcommand logic to the library
Introduce the Driver which does the orchestration that was previously done in
the main.rs of cargo-minver. It returns an Analysis struct, which encapsulates
the list of features. Some more logic will be added to the impl of this
struct, e.g. returning the actual minimum version, along other stuff.

This step enables integration testing of cargo-minver as we will be able to
evaluate the result of the analysis programmatically.