Base braced empty structs detection on HIR + track enums
The previous implementation was erroneous because we can't really evaluate for
pattern and expression positions if the related struct has no fields as we can't
derive that information from how the struct is used. We were also missing enum
variants that are structs.

With the new implementation we use the resolved path to the struct to check the
number of fields in the definition.

The test is now more exhaustive.